logical restructuring fixes for new workflow
changed collateci to be 'bashlike' -> takes a list
of input folders, and outputs the result. going
to alter all the collate scripts to work like this
at some point. problem is that it needs at least
four arguments - finding it hard to do defaults
with arrays

changed collatepadloc to omit the '_padloc' at end
of folder name in compilation outputs

reworked Snakefile to actually work